move PLUTO integration gem to a more standard package layout and add tests for notification
