+++ CDS/Ruby/PLUTO/Rakefile
@@ -1,0 +1,16 @@
+require 'rake/testtask'
+require 'rubygems/package_task'
+begin
+  require 'rspec/core/rake_task'
+  RSpec::Core::RakeTask.new(:spec) do |t, task_args|
+    t.rspec_opts = "--require spec_helper"
+  end
+rescue LoadError
+end
+
+desc "Run tests"
+task :default => :spec
+
+task :build do
+  sh "gem build plutolib.gemspec"
+end
