+++ CDS/Ruby/PLUTO/lib/PLUTO/Entity.rb
