+++ CDS/Ruby/PLUTO/spec/spec_helper.rb
@@ -1,0 +1,3 @@
+require 'webmock/rspec'
+
+WebMock.disable_net_connect!(allow_localhost: false)
