+++ dockerbuild/ruby_setup.sh
@@ -9,4 +9,5 @@ DEBIAN_FRONTEND=noninteractive apt-add-repository -y ppa:brightbox/ruby-ng
 apt-get -y update
 apt-get -y install ruby2.6 ruby2.6-dev ruby-switch
 ruby-switch --set ruby2.6
-gem install awesome_print trollop sentry-raven aws-sdk-v1 aws-sdk-core aws-sdk-resources google-api-client:'<0.9' launchy thin rest-client certifi sentry-raven xmp mail exifr elasticsearch
+gem install awesome_print trollop sentry-raven aws-sdk-v1 aws-sdk-core aws-sdk-resources google-api-client:'<0.9' \
+    launchy thin rest-client certifi sentry-raven xmp mail exifr elasticsearch rake rspec webmock
